Added cardboard framework instead of multiple pop-up windows
