+++ nbproject/private/private.xml
@@ -12,10 +12,7 @@
     </editor-bookmarks>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/Users/ntko/Documents/NetBeansProjects/AutomateYourJob/src/automateyourjob/templates/Framework.java</file>
-            <file>file:/C:/Users/ntko/Documents/NetBeansProjects/AutomateYourJob/src/automateyourjob/templates/EmploymentTemplateModifier.java</file>
             <file>file:/C:/Users/ntko/Documents/NetBeansProjects/AutomateYourJob/src/automateyourjob/AutomateYourJob.java</file>
-            <file>file:/C:/Users/ntko/Documents/NetBeansProjects/AutomateYourJob/src/automateyourjob/templates/ReadTemplate.java</file>
         </group>
     </open-files>
 </project-private>
