+++ src/automateyourjob/templates/AutomateYourJobUI.java
@@ -1,0 +1,493 @@
+/*
+ * Copyright (C) 2016 ntko
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package automateyourjob.templates;
+
+import java.awt.CardLayout;
+import java.io.BufferedWriter;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
+/**
+ *
+ * @author ntko
+ */
+public class AutomateYourJobUI extends javax.swing.JFrame {
+
+    /**
+     * Creates new form NewJFrame
+     */
+    public AutomateYourJobUI() {
+        initComponents();
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        jScrollPane1 = new javax.swing.JScrollPane();
+        jTextPane1 = new javax.swing.JTextPane();
+        mainPanel = new javax.swing.JPanel();
+        homePanel = new javax.swing.JPanel();
+        employmentEdit = new javax.swing.JPanel();
+        jLabel1 = new javax.swing.JLabel();
+        jLabel2 = new javax.swing.JLabel();
+        jReferee = new javax.swing.JTextField();
+        jTitle = new javax.swing.JTextField();
+        ArrayList<String> userSetText = null;
+        try {
+            ReadTemplate r = new ReadTemplate();
+            String fileName = "automateyourjob\\templates\\employment.template";
+
+            userSetText = r.ReadTemplate(fileName);
+        }catch (IOException ex) {
+            Logger.getLogger(AutomateYourJobUI.class.getName()).log(Level.SEVERE, null, ex);
+        }
+        jMainText = new javax.swing.JTextField();
+        jDate1 = new javax.swing.JTextField();
+        jTo = new javax.swing.JTextField();
+        jDate2 = new javax.swing.JTextField();
+        jPositionText = new javax.swing.JTextField();
+        jPosition = new javax.swing.JTextField();
+        jSalaryText = new javax.swing.JTextField();
+        jSalary = new javax.swing.JTextField();
+        jScrollPane2 = new javax.swing.JScrollPane();
+        jInfoText = new javax.swing.JTextArea();
+        jSaveButton = new javax.swing.JButton();
+        jDefaultButton = new javax.swing.JButton();
+        jMenuBar1 = new javax.swing.JMenuBar();
+        jMenu1 = new javax.swing.JMenu();
+        jMenuItem1 = new javax.swing.JMenuItem();
+        jMenuItem3 = new javax.swing.JMenuItem();
+        jMenuItem4 = new javax.swing.JMenuItem();
+        jMenuItem5 = new javax.swing.JMenuItem();
+        jMenu2 = new javax.swing.JMenu();
+        jMenuItem2 = new javax.swing.JMenuItem();
+
+        jScrollPane1.setHorizontalScrollBarPolicy(javax.swing.ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER);
+        jScrollPane1.setVerticalScrollBarPolicy(javax.swing.ScrollPaneConstants.VERTICAL_SCROLLBAR_NEVER);
+        jScrollPane1.setViewportView(jTextPane1);
+
+        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
+        setAlwaysOnTop(true);
+        setResizable(false);
+        addContainerListener(new java.awt.event.ContainerAdapter() {
+            public void componentAdded(java.awt.event.ContainerEvent evt) {
+                formComponentAdded(evt);
+            }
+        });
+        addMouseMotionListener(new java.awt.event.MouseMotionAdapter() {
+            public void mouseDragged(java.awt.event.MouseEvent evt) {
+                formMouseDragged(evt);
+            }
+        });
+        getContentPane().setLayout(new java.awt.CardLayout());
+
+        mainPanel.setLayout(new java.awt.CardLayout());
+
+        javax.swing.GroupLayout homePanelLayout = new javax.swing.GroupLayout(homePanel);
+        homePanel.setLayout(homePanelLayout);
+        homePanelLayout.setHorizontalGroup(
+            homePanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 663, Short.MAX_VALUE)
+        );
+        homePanelLayout.setVerticalGroup(
+            homePanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 181, Short.MAX_VALUE)
+        );
+
+        mainPanel.add(homePanel, "homePanel");
+
+        employmentEdit.setInheritsPopupMenu(true);
+
+        jLabel1.setText("Separate parts are modifiable");
+
+        jLabel2.setText("These fields will not appear if the data hasn't been entered");
+
+        jReferee.setEditable(false);
+        jReferee.setFont(new java.awt.Font("Dialog", 0, 10)); // NOI18N
+        jReferee.setText("<Referee>");
+
+        jTitle.setEditable(false);
+        jTitle.setFont(new java.awt.Font("Dialog", 0, 10)); // NOI18N
+        jTitle.setText("(<Title>)");
+        jTitle.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jTitleActionPerformed(evt);
+            }
+        });
+
+        jMainText.setFont(new java.awt.Font("Dialog", 0, 10)); // NOI18N
+        jMainText.setText(userSetText.get(0));
+
+        jDate1.setEditable(false);
+        jDate1.setFont(new java.awt.Font("Dialog", 0, 10)); // NOI18N
+        jDate1.setText("<Date 1>");
+
+        jTo.setFont(new java.awt.Font("Dialog", 0, 10)); // NOI18N
+        jTo.setText(userSetText.get(1));
+
+        jDate2.setEditable(false);
+        jDate2.setFont(new java.awt.Font("Dialog", 0, 10)); // NOI18N
+        jDate2.setText("<Date 2>");
+
+        jPositionText.setFont(new java.awt.Font("Dialog", 0, 10)); // NOI18N
+        jPositionText.setText(userSetText.get(2));
+
+        jPosition.setEditable(false);
+        jPosition.setFont(new java.awt.Font("Dialog", 0, 10)); // NOI18N
+        jPosition.setText("<Position>");
+
+        jSalaryText.setFont(new java.awt.Font("Dialog", 0, 10)); // NOI18N
+        jSalaryText.setText(userSetText.get(3));
+
+        jSalary.setEditable(false);
+        jSalary.setFont(new java.awt.Font("Dialog", 0, 10)); // NOI18N
+        jSalary.setText("<salary>");
+
+        jInfoText.setColumns(20);
+        jInfoText.setFont(new java.awt.Font("Dialog", 0, 10)); // NOI18N
+        jInfoText.setRows(5);
+        jInfoText.setText(userSetText.get(4));
+        jInfoText.addContainerListener(new java.awt.event.ContainerAdapter() {
+            public void componentAdded(java.awt.event.ContainerEvent evt) {
+                jInfoTextComponentAdded(evt);
+            }
+        });
+        jScrollPane2.setViewportView(jInfoText);
+
+        jSaveButton.setText("Save");
+        jSaveButton.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseReleased(java.awt.event.MouseEvent evt) {
+                jSaveButtonMouseReleased(evt);
+            }
+        });
+        jSaveButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jSaveButtonActionPerformed(evt);
+            }
+        });
+
+        jDefaultButton.setText("Default");
+        jDefaultButton.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseReleased(java.awt.event.MouseEvent evt) {
+                jDefaultButtonMouseReleased(evt);
+            }
+        });
+        jDefaultButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jDefaultButtonActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout employmentEditLayout = new javax.swing.GroupLayout(employmentEdit);
+        employmentEdit.setLayout(employmentEditLayout);
+        employmentEditLayout.setHorizontalGroup(
+            employmentEditLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(employmentEditLayout.createSequentialGroup()
+                .addGap(22, 22, 22)
+                .addGroup(employmentEditLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                    .addComponent(jScrollPane2)
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, employmentEditLayout.createSequentialGroup()
+                        .addComponent(jReferee, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jTitle, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jMainText, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jDate1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jTo, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jDate2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jPositionText, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jPosition, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jSalaryText, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jSalary, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addGroup(employmentEditLayout.createSequentialGroup()
+                        .addComponent(jLabel1)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(jLabel2))
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, employmentEditLayout.createSequentialGroup()
+                        .addComponent(jSaveButton)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jDefaultButton)))
+                .addContainerGap(22, Short.MAX_VALUE))
+        );
+        employmentEditLayout.setVerticalGroup(
+            employmentEditLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(employmentEditLayout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(employmentEditLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel2)
+                    .addComponent(jLabel1))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(employmentEditLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jReferee, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jTitle, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jMainText, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jDate1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jTo, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jDate2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jPositionText, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jPosition, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jSalaryText, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jSalary, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jScrollPane2, javax.swing.GroupLayout.PREFERRED_SIZE, 81, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(employmentEditLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jSaveButton)
+                    .addComponent(jDefaultButton))
+                .addContainerGap(12, Short.MAX_VALUE))
+        );
+
+        mainPanel.add(employmentEdit, "employmentEdit");
+
+        getContentPane().add(mainPanel, "mainPanel");
+
+        jMenuBar1.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseReleased(java.awt.event.MouseEvent evt) {
+                jMenuBar1MouseReleased(evt);
+            }
+        });
+
+        jMenu1.setText("File");
+
+        jMenuItem1.setAccelerator(javax.swing.KeyStroke.getKeyStroke(java.awt.event.KeyEvent.VK_H, java.awt.event.InputEvent.ALT_MASK));
+        jMenuItem1.setText("Home");
+        jMenuItem1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jMenuItem1ActionPerformed(evt);
+            }
+        });
+        jMenu1.add(jMenuItem1);
+
+        jMenuItem3.setAccelerator(javax.swing.KeyStroke.getKeyStroke(java.awt.event.KeyEvent.VK_P, java.awt.event.InputEvent.ALT_MASK));
+        jMenuItem3.setText("Awaiting Publishing");
+        jMenu1.add(jMenuItem3);
+
+        jMenuItem4.setAccelerator(javax.swing.KeyStroke.getKeyStroke(java.awt.event.KeyEvent.VK_C, java.awt.event.InputEvent.ALT_MASK));
+        jMenuItem4.setText("Candidate Comms");
+        jMenu1.add(jMenuItem4);
+
+        jMenuItem5.setAccelerator(javax.swing.KeyStroke.getKeyStroke(java.awt.event.KeyEvent.VK_O, java.awt.event.InputEvent.ALT_MASK));
+        jMenuItem5.setText("Other");
+        jMenu1.add(jMenuItem5);
+
+        jMenuBar1.add(jMenu1);
+
+        jMenu2.setText("Edit Template");
+        jMenu2.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseReleased(java.awt.event.MouseEvent evt) {
+                jMenu2MouseReleased(evt);
+            }
+        });
+
+        jMenuItem2.setText("Employment");
+        jMenuItem2.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                jMenuItem2MouseClicked(evt);
+            }
+            public void mouseReleased(java.awt.event.MouseEvent evt) {
+                jMenuItem2MouseReleased(evt);
+            }
+        });
+        jMenuItem2.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jMenuItem2ActionPerformed(evt);
+            }
+        });
+        jMenu2.add(jMenuItem2);
+
+        jMenuBar1.add(jMenu2);
+
+        setJMenuBar(jMenuBar1);
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void formMouseDragged(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_formMouseDragged
+        // TODO add your handling code here:
+    }//GEN-LAST:event_formMouseDragged
+
+    private void jMenuItem2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jMenuItem2ActionPerformed
+    CardLayout card = (CardLayout)mainPanel.getLayout();
+    card.show(mainPanel, "employmentEdit");
+    }//GEN-LAST:event_jMenuItem2ActionPerformed
+
+    private void formComponentAdded(java.awt.event.ContainerEvent evt) {//GEN-FIRST:event_formComponentAdded
+        // TODO add your handling code here:
+    }//GEN-LAST:event_formComponentAdded
+
+    private void jMenuBar1MouseReleased(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jMenuBar1MouseReleased
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jMenuBar1MouseReleased
+
+    private void jMenu2MouseReleased(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jMenu2MouseReleased
+
+    }//GEN-LAST:event_jMenu2MouseReleased
+
+    private void jMenuItem2MouseReleased(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jMenuItem2MouseReleased
+
+    }//GEN-LAST:event_jMenuItem2MouseReleased
+
+    private void jTitleActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jTitleActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jTitleActionPerformed
+
+    private void jInfoTextComponentAdded(java.awt.event.ContainerEvent evt) {//GEN-FIRST:event_jInfoTextComponentAdded
+
+    }//GEN-LAST:event_jInfoTextComponentAdded
+
+    private void jSaveButtonMouseReleased(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jSaveButtonMouseReleased
+
+    }//GEN-LAST:event_jSaveButtonMouseReleased
+
+    private void jSaveButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jSaveButtonActionPerformed
+            try {
+            String[] employment;
+            employment = new String[5];
+            employment[0] = jMainText.getText();
+            employment[1] = jTo.getText();
+            employment[2] = jPositionText.getText();
+            employment[3] = jSalaryText.getText();
+            employment[4] = jInfoText.getText();
+            int nbrEmployment = employment.length;
+
+            BufferedWriter writer = null;
+            writer = new BufferedWriter(new FileWriter("src\\automateyourjob\\templates\\employment.template"));
+            for(int i = 0; i < nbrEmployment; i++)
+            {
+                writer.write(employment[i]);
+                writer.append('¦');
+            }
+            writer.flush();
+            writer.close();
+
+        } catch (IOException ex) {
+            Logger.getLogger(AutomateYourJobUI.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    }//GEN-LAST:event_jSaveButtonActionPerformed
+
+    private void jMenuItem2MouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jMenuItem2MouseClicked
+    }//GEN-LAST:event_jMenuItem2MouseClicked
+
+    private void jMenuItem1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jMenuItem1ActionPerformed
+    CardLayout card = (CardLayout)mainPanel.getLayout();
+    card.show(mainPanel, "homePanel");
+    }//GEN-LAST:event_jMenuItem1ActionPerformed
+
+    private void jDefaultButtonMouseReleased(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jDefaultButtonMouseReleased
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jDefaultButtonMouseReleased
+
+    private void jDefaultButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jDefaultButtonActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jDefaultButtonActionPerformed
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String args[]) {
+        /* Set the Nimbus look and feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
+                }
+            }
+        } catch (ClassNotFoundException ex) {
+            java.util.logging.Logger.getLogger(AutomateYourJobUI.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            java.util.logging.Logger.getLogger(AutomateYourJobUI.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            java.util.logging.Logger.getLogger(AutomateYourJobUI.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(AutomateYourJobUI.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        }
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+
+        /* Create and display the form */
+        java.awt.EventQueue.invokeLater(() -> {
+            new AutomateYourJobUI().setVisible(true);
+        });
+    }
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JPanel employmentEdit;
+    private javax.swing.JPanel homePanel;
+    private javax.swing.JTextField jDate1;
+    private javax.swing.JTextField jDate2;
+    private javax.swing.JButton jDefaultButton;
+    private javax.swing.JTextArea jInfoText;
+    private javax.swing.JLabel jLabel1;
+    private javax.swing.JLabel jLabel2;
+    private javax.swing.JTextField jMainText;
+    private javax.swing.JMenu jMenu1;
+    private javax.swing.JMenu jMenu2;
+    private javax.swing.JMenuBar jMenuBar1;
+    private javax.swing.JMenuItem jMenuItem1;
+    private javax.swing.JMenuItem jMenuItem2;
+    private javax.swing.JMenuItem jMenuItem3;
+    private javax.swing.JMenuItem jMenuItem4;
+    private javax.swing.JMenuItem jMenuItem5;
+    private javax.swing.JTextField jPosition;
+    private javax.swing.JTextField jPositionText;
+    private javax.swing.JTextField jReferee;
+    private javax.swing.JTextField jSalary;
+    private javax.swing.JTextField jSalaryText;
+    private javax.swing.JButton jSaveButton;
+    private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JScrollPane jScrollPane2;
+    private javax.swing.JTextPane jTextPane1;
+    private javax.swing.JTextField jTitle;
+    private javax.swing.JTextField jTo;
+    private javax.swing.JPanel mainPanel;
+    // End of variables declaration//GEN-END:variables
+}
