+++ src/automateyourjob/templates/ReadTemplate.java
@@ -31,15 +31,12 @@ import java.util.Arrays;
  */
 public class ReadTemplate {
 
-    public ReadTemplate(String srcautomateyourjobtemplatesemploymenttemp) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    void Reader(String fileName) throws IOException {
+    public ArrayList<String> ReadTemplate(String fileName) throws IOException {
 
             URL url = Resources.getResource(fileName);
             String text = Resources.toString(url, Charsets.UTF_8);
-            ArrayList<String> textArray = new ArrayList<>(Arrays.asList(text.split("|")));
-            System.out.print(textArray);
+            ArrayList<String> textArray = new ArrayList<>(Arrays.asList(text.split("¦")));
+            return textArray;
         }
-    }
+}
+
