+++ src/automateyourjob/templates/employment.template
@@ -1,5 +1,5 @@
-has advised that the Candidate was employed from|to|in the position of|with a salary of|The Referee described the Candidate  as
+has advised that the Candidate was employed from¦to¦in the position of¦with a salary of¦The Referee described the Candidate  as
 
 They further advised that the Candidate was honest and showed integrity, and the Candidate's health and disciplinary record were clean.
 
-The reason for leaving stated was XXX and the Company would re-employ the Candidate.||
+The reason for leaving stated was XXX and the Company would re-employ the Candidate.¦
